DownloadBlocks visitor
