+++ src/Stratis.Bitcoin/Consensus/Visitors/ConsensusVisitorResult.cs
@@ -1,0 +1,6 @@
+﻿namespace Stratis.Bitcoin.Consensus.Visitors
+{
+    public sealed class ConsensusVisitorResult
+    {
+    }
+}
