+++ src/Stratis.Bitcoin/Consensus/Visitors/IConsensusVisitor.cs
@@ -1,0 +1,7 @@
+﻿namespace Stratis.Bitcoin.Consensus.Visitors
+{
+    public interface IConsensusVisitor
+    {
+        ConsensusVisitorResult Visit(ConsensusManager consensusManager);
+    }
+}
